LPS-55159 Rename module
