--- modules/test/cobertura-agent/bnd.bnd
@@ -1,7 +1,0 @@
-Bundle-Name: Liferay Cobertura Agent
-Bundle-SymbolicName: com.liferay.cobertura.agent
-Bundle-Version: 1.0.0
-Can-Redefine-Classes: true
-Can-Retransform-Classes: true
-Include-Resource: classes
-Premain-Class: com.liferay.cobertura.agent.InstrumentationAgent
