+++ modules/test/whip/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Whip
+Bundle-SymbolicName: com.liferay.whip
+Bundle-Version: 1.0.0
+Can-Redefine-Classes: true
+Can-Retransform-Classes: true
+Include-Resource: classes
+Premain-Class: com.liferay.whip.agent.InstrumentationAgent
